% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Download_Refs}
\alias{Download_Refs}
\title{download_refs}
\usage{
Download_Refs(ID_file, assembly_location, outdir)
}
\arguments{
\item{ID_file}{the location of the file with list of organisms generated by parse_report()}

\item{assembly_location}{the location of the refseq metadata files you downloaded using get_metadata()}

\item{outdir}{The filepath you want to save your reference genomes to}
}
\value{
downloades reference genomes using lists of organism names from parse_report() files
}
\description{
downloads reference genomes using species lists generated by the parse_report() function. It will only download full, complete genomes from the refseq database.

License: MIT + file LICENSE
}
\examples{
\dontrun{
path = "~/Test_Data/Lactobacillaceae.txt"
location = "~/Full_test"
Download_Refs(path, assembly_location = location, outdir = location)}

}
