% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(kingdom, outdir)
}
\arguments{
\item{kingdom}{the kingdoms that the refseq genomes you want to obtain for downstream analysis
come from. These can be any of these options:  c("archaea",
"fungi",
"bacteria",
"invertebrate",
"plant",
"protozoa",
"vertebrate_mammalian",
"vertebrate_other",
"viral")}

\item{outdir}{The filepath you want to save your metadata files to}
}
\value{
saves refseq metadata to your machine - necessary to download genomes
}
\description{
downloads refseq metadata - you need to run this once after downloading this package for every kingdom
you are interested in (e.g. bacteria, fungi).

License: MIT + file LICENSE
}
\examples{
\dontrun{
kings <- "fungi"
file = "~/Full_test"
get_metadata(kings, file)}

}
