% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Parse_report}
\alias{Parse_report}
\title{Krakextract}
\usage{
Parse_report(
  file,
  filetype,
  sci_names,
  rank_method,
  rank_no = 5,
  min_reads = 0,
  outdir
)
}
\arguments{
\item{file}{This is the file path to your Kraken report file}

\item{filetype}{This is the type of file, only "Kraken" works at this point}

\item{sci_names}{This is a vector of the form c("Lactobacillaceae", "Streptococcaceae") ONLY FAMILY NAMES WORK}

\item{rank_method}{Sort the kraken hits by percentage or number of reads args = "reads" or "percent"}

\item{rank_no}{The top x number of species (default is 5)}

\item{min_reads}{The minimum number of reads for a species to be included (default is 0)}

\item{outdir}{The filepath you want to save your text files to}
}
\value{
Text files with NCBI IDs for each species of interest above thresholds in your kraken output
}
\description{
This package extracts species of interest from Kraken
reports allowing users to download and align
their reference genomes from NCBI

License: MIT + file LICENSE
}
\examples{
\dontrun{
families = c("Lactobacillaceae", "Aerococcaceae")
path_out = "man/Example_Data/"

Parse_report("man/Example_Data/Example.kreport",
filetype="Kraken",
sci_names=families,
rank_method="reads",
rank_no=10,
outdir=path_out)}

}
